requirements/base.txt
Django==6.0
djangorestframework
psycopg2-binary
python-dotenv
celery
redis
django-cors-headers
